requirements.txt
colorama
python-nmap
os
webbrowser
